requirements.txt
altgraph==0.17.4
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
python-dotenv==1.1.1
pywin32-ctypes==0.2.3
qrcode==8.2
Werkzeug==3.1.3
